requirements.txt
# 核心依赖
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
pillow>=9.5.0
numpy>=1.24.0

# 工具库
tqdm>=4.65.0
einops>=0.6.1

# 训练和可视化
tensorboard>=2.13.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

